requirement.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
certifi==2017.11.5
chardet==3.0.4
cssselect==1.0.1
decorator==4.1.2
enum34==1.1.6
gprof2dot==2017.9.19
graphviz==0.8.1
idna==2.6
ipython==5.5.0
ipython-genutils==0.2.0
line-profiler==2.0
lxml==4.1.1
memory-profiler==0.50.0
olefile==0.44
pathlib2==2.3.0
pexpect==4.3.0
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
psutil==5.4.1
ptyprocess==0.5.2
Pygments==2.2.0
pytesseract==0.1.7
readability-lxml==0.6.2
requests==2.18.4
scandir==1.6
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
